Fix saving disable_refresh setting
The sanitize_settings() function is called twice when the option does not exist. The current validation for the 'disable_refresh' setting causes the setting to always be updated to `false` if the setting has never been saved before. This update adds an additional check that explicitly verifies if the setting is true-ish before disabling.

Related trac ticket: https://core.trac.wordpress.org/ticket/21989